Split gigantic ui_template.go to multiple files for easier maintainability
