+++ .github/workflows/nightly.yaml
@@ -24,7 +24,7 @@ jobs:
       runs-on: ubuntu-latest
       os-name: linux
       timeout-minutes: 120
-      label-filter: "-ginkgo.label-filter='!(smoke, upgrade)'"
+      label-filter: "--label-filter='!(smoke, upgrade)'"
       kubectl-version: "v1.23.3"
       login_user_type: "oidc"
       git-provider: gitlab
@@ -56,7 +56,7 @@ jobs:
       runs-on: ubuntu-latest
       os-name: linux
       timeout-minutes: 90
-      label-filter: "-ginkgo.label-filter='!(capd, upgrade, kind-gitops-cluster, leaf-application, leaf-policy, leaf-violation)'"
+      label-filter: "--label-filter='!(capd, upgrade, kind-gitops-cluster, leaf-application, leaf-policy, leaf-violation)'"
       kubectl-version: "v1.23.6"
       login_user_type: "oidc"
       git-provider: github
@@ -87,7 +87,7 @@ jobs:
       runs-on: ubuntu-latest
       os-name: linux
       timeout-minutes: 90
-      label-filter: "-ginkgo.label-filter='!(upgrade, capd, kind-gitops-cluster, tenant, leaf-application, leaf-policy, leaf-violation)'"
+      label-filter: "--label-filter='!(upgrade, capd, kind-gitops-cluster, tenant, leaf-application, leaf-policy, leaf-violation)'"
       kubectl-version: "v1.23.3"
       login_user_type: "cluster-user"
       git-provider: gitlab
