+++ test/utils/data/templates/application/multi-create-request-helmrelease-template.yaml
@@ -1,0 +1,108 @@
+apiVersion: templates.weave.works/v1alpha1
+kind: GitOpsTemplate
+metadata:
+  name: mvp-helmrelease-template
+  namespace: default
+  labels:
+    weave.works/template-type: application
+  annotations:
+    templates.weave.works/cost-estimation-enabled: "false"
+    templates.weave.works/profiles-enabled: "false"
+    templates.weave.works/credentials-enabled: "false"
+    templates.weave.works/kustomizations-enabled: "false"
+    templates.weave.works/delimiters: "${{,}}"
+spec:
+  renderType: templating
+  description: This is a template for deploying application as helemrelease using a Flux Kustomization
+  params:
+    - name: RESOURCE_NAME
+      required: true
+      description: Source that contains the Kubernetes manifests
+    - name: NAMESPACE
+      description: Namespace to create the source in
+      default: default    
+    - name: URL
+      description: URL of the git repository
+    - name: INTERVAL
+      description: How often to check for updates
+      default: 30s
+    - name: CHART_NAME
+      required: true
+      description: Helm chart name
+    - name: CHART_VERSION
+      required: true
+      description:  Helm chart version, accepts a semver range
+    - name: TARGET_NAMESPACE
+      required: true
+      default: default
+      description: Namespace of all Kustomization objects
+    - name: VALUES
+      required: false
+      default: nil
+      description: "Values for the helm chart,e.g. {ui: {message: hello}}"
+    - name: HOST_NAME
+      required: true
+      description: Host name running backend services for the deployment(s)
+    - name: SERVICE_NAME
+      required: true
+      description: HTTP service name exposed by ingress-nginx
+    - name: SERVICE_PORT
+      required: true
+      description: ClusterIP service port
+  resourcetemplates:
+    - apiVersion: source.toolkit.fluxcd.io/v1beta1
+      kind: HelmRepository
+      metadata:
+        name: ${{ .params.RESOURCE_NAME }}
+        namespace: ${{ .params.NAMESPACE }}
+        annotations:
+          templates.weave.works/create-request: ""
+      spec:
+        interval: ${{ .params.INTERVAL }}
+        url: ${{ .params.URL }}
+    - apiVersion: helm.toolkit.fluxcd.io/v2beta1
+      kind: HelmRelease
+      metadata:
+        name: ${{ .params.RESOURCE_NAME }}
+        namespace: ${{ .params.NAMESPACE }}
+        annotations:
+          templates.weave.works/create-request: ""
+      spec:
+        interval: 30s
+        chart:
+          spec:
+            chart: ${{ .params.CHART_NAME }}
+            version: ${{ .params.CHART_VERSION }}
+            sourceRef:
+              apiVersion: source.toolkit.fluxcd.io/v1beta2
+              kind: HelmRepository
+              name: ${{ .params.RESOURCE_NAME }}
+              namespace:  ${{ .params.NAMESPACE }}
+        targetNamespace: ${{ .params.TARGET_NAMESPACE }}
+        values: ${{ .params.VALUES }}
+    - apiVersion: networking.k8s.io/v1
+      kind: Ingress
+      metadata:
+        annotations:
+          cert-manager.io/issuer: selfsigned-issuer
+          kubernetes.io/ingress.class: "nginx"
+          nginx.ingress.kubernetes.io/rewrite-target: /$1
+        name: ${{ .params.RESOURCE_NAME }}
+        namespace: ${{ .params.TARGET_NAMESPACE }}
+      spec:
+        rules:
+        - host: ${{ .params.HOST_NAME}}
+          http:
+            paths:
+            - pathType: ImplementationSpecific
+              path: /?(.*)
+              backend:
+                service:
+                  name: ${{ .params.SERVICE_NAME}}
+                  port:
+                    number: ${{ .params.SERVICE_PORT}}
+        tls:
+        - hosts:
+          - ${{ .params.HOST_NAME}}
+          secretName: weave-gitops-tls
+
