+++ test/utils/data/templates/cluster/aws/cluster-template-eks-fargate.yaml
@@ -3,6 +3,10 @@ kind: CAPITemplate
 metadata:
   name: capa-cluster-template-eks-fargate-{{.Count}}
   namespace: default
+  labels:
+    weave.works/template-type: cluster
+  annotations:
+    templates.weave.works/inject-prune-annotation: "true"    
 spec:
   description: This is eks fargate template-{{.Count}}
   params:
