+++ test/utils/data/templates/cluster/azure/cluster-template-e2e.yaml
@@ -3,7 +3,10 @@ kind: CAPITemplate
 metadata:
   name: capz-cluster-template-{{.Count}}
   namespace: default
+  labels:
+    weave.works/template-type: cluster
   annotations:
+    templates.weave.works/inject-prune-annotation: "true"
     templates.weave.works/credentials-enabled: "true"
 spec:
   description: This is Azure capi quick start template-{{.Count}}
